server/pyproject.toml
[project]
name = "homepage-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.132.0",
    "uvicorn[standard]>=0.41.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "python-jose>=3.5.0",
    "passlib>=1.7.4",
    "bcrypt>=5.0.0",
    "python-multipart>=0.0.22",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.1",
    "pytest",
    "ruff",
    "basedpyright",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]

[tool.basedpyright]
exclude = [
    "tests",
]
